requirements.txt
alembic==1.14.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.6
MarkupSafe==3.0.1
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scikit-fuzzy==0.5.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
SQLAlchemy==2.0.35
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.4
